Migrate new EPC vis into EPC page
