+++ prospector-frontend/src/epc/index.html
@@ -1,6 +1,6 @@
 ---
 layout: standalone
-title: Equity Priority Communities
+title: Equity Priority Communities Vis 2
 thumbnail: scrnshot.png
 css:
   - '/assets/css/standalone.css'
