--- prospector-frontend/src/epc2/index.html
@@ -1,11 +1,0 @@
----
-layout: standalone
-title: Equity Priority Communities Vis 2
-thumbnail: scrnshot.png
-css:
-  - '/assets/css/standalone.css'
----
-
-{% include epc2.html %}
-
-<script type="application/javascript" src="/bundles/epc2.js"></script>
